+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 - Add `Block` to allow for creation of arbitrary blocks. (#276)
 - Add functionality to fetch, search or remove a block using the ORCA name of the block. (#276)
 - Added `GbwResults.get_structure()`, `GbwResults.from_gbw_file()`, and `Output.get_structure_from_gbw()` to obtain a `Structure` from a gbw file or gbw JSON (#280).
+- Added `Output.get_timings()` to access the timings of the calculation steps (#284).
 
 ### Changed
 - Refactored methods from Runner into BaseRunner (#193)
@@ -53,6 +54,7 @@
 - Fixed a typo in the badge for the OPI paper (#222).
 - Fixed `Structure.nelectrons` for structures containing ghost atoms (#268).
 - Fixed `_orca_environment()` which now makes changes to `os.environ` in-place without breaking any reference to that dict (#279).
+- Negative calculation timings occasionally reported by ORCA are now clamped to zero instead of raising a `ValidationError` (#284).
 
 ## [2.0.0] - 2026-02-10